+++ idle-backend/src/main/java/com/idle/mapper/GoodsMapper.java
@@ -14,30 +14,30 @@ import java.util.List;
 public interface GoodsMapper {
 
     @Select("select * from db_goods where id = #{id}")
-    List<Goods> findGoodsById(Integer id);
+    List<Goods> findGoodsById(Long id);
 
     @Insert("insert into db_goods (id, publisher_id, product_name, price, category, introduce, picture, create_time)" +
             " values (#{id}, #{publisher_id}, #{product_name}, #{price}, #{category}, #{introduce}," +
             " #{picture}, #{create_time})")
-    int createGoods(Integer id, Integer publisher_id, String product_name, BigDecimal price, String category,
+    int createGoods(Long id, Long publisher_id, String product_name, BigDecimal price, String category,
                     String introduce, String picture, String create_time);
 
     @Update("update db_goods set product_name = #{product_name}, price = #{price}, category = #{price}, introduce = " +
             "#{introduce}, picture = #{picture} where id = #{id}")
-    int updateGoods(Integer id,String product_name, BigDecimal price, String category, String introduce, String picture);
+    int updateGoods(Long id,String product_name, BigDecimal price, String category, String introduce, String picture);
 
     @Update("update db_goods set status = '0' where id = #{id}")
-    int updateStatus(Integer id);
+    int updateStatus(Long id);
 
     @Select("select * from db_goods")
     List<Goods> getAllGoods();
 
     @Select("select * from db_goods join db_account on db_account.id = db_goods.publisher_id where" +
             " db_goods.publisher_id = #{publisher_id}")
-    List<Goods> getGoodsByPublisherId(Integer publisher_id);
+    List<Goods> getGoodsByPublisherId(Long publisher_id);
 
     @Select("select * from db_account where db_account.id = #{publisher_id}")
-    AccountUser getUserByPublisherId(Integer publisher_id);
+    AccountUser getUserByPublisherId(Long publisher_id);
 
     @Select("select * from db_goods where category = #{category}")
     List<Goods> getGoodsByCategory(String category);
@@ -46,5 +46,5 @@ public interface GoodsMapper {
             "OR db_goods.introduce like CONCAT('%', #{search}, '%')")
     List<Goods> getGoodsByProductNameOrIntroduce(String search);
     @Select("select price from db_goods where id = #{id}")
-    String getPriceById(Integer id);
+    String getPriceById(Long id);
 }
